requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
sqlalchemy==2.0.23
pillow==10.1.0